Fix dm related tests
